Changes to be committed: 	modified:   app/Http/Controllers/TransactionController.php 	modified:   app/Models/Transaction.php 	deleted:    app/Policies/TransactionPolicy.php 	modified:   app/Providers/AppServiceProvider.php 	modified:   database/migrations/2014_10_12_000000_create_users_table.php 	modified:   database/migrations/2014_10_12_100000_create_password_resets_table.php 	modified:   database/migrations/2019_08_19_000000_create_failed_jobs_table.php 	modified:   database/migrations/2019_12_14_000001_create_personal_access_tokens_table.php 	modified:   database/migrations/2023_03_06_121814_create_transactions_table.php 	modified:   database/migrations/2023_03_06_172611_create_categories_table.php 	modified:   database/migrations/2023_03_06_173718_add_column_to_table.php 	modified:   phpunit.xml 	modified:   resources/views/home.blade.php 	modified:   resources/views/layouts/app.blade.php 	new file:   resources/views/transactions/_form.blade.php 	new file:   resources/views/transactions/create.blade.php 	new file:   resources/views/transactions/edit.blade.php 	modified:   resources/views/transactions/index.blade.php 	modified:   routes/web.php 	modified:   tests/Feature/CreateTransactionsTest.php 	new file:   tests/Feature/DeleteTransactionTest.php 	deleted:    tests/Feature/ExampleTest.php 	new file:   tests/Feature/UpdateTransactionsTest.php 	modified:   tests/Feature/ViewTransactionsTest.php 	modified:   tests/TestCase.php

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -3,6 +3,8 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\App;
+
 
 class CreateUsersTable extends Migration
 {
@@ -13,6 +15,11 @@ class CreateUsersTable extends Migration
      */
     public function up()
     {
+        if (App::environment() === 'production') {
+            // Prevent running migrations in production environment
+            return;
+        }
+
         Schema::create('users', function (Blueprint $table) {
             $table->id();
             $table->string('name');
@@ -31,6 +38,10 @@ class CreateUsersTable extends Migration
      */
     public function down()
     {
+        if (App::environment() === 'production') {
+            // Prevent running migrations in production environment
+            return;
+        }
         Schema::dropIfExists('users');
     }
 }
