+++ database/migrations/2014_10_12_100000_create_password_resets_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\App;
 
 return new class extends Migration
 {
@@ -13,6 +14,11 @@ return new class extends Migration
      */
     public function up()
     {
+        if (App::environment() === 'production') {
+            // Prevent running migrations in production environment
+            return;
+        }
+
         Schema::create('password_resets', function (Blueprint $table) {
             $table->string('email')->index();
             $table->string('token');
@@ -27,6 +33,11 @@ return new class extends Migration
      */
     public function down()
     {
+        if (App::environment() === 'production') {
+            // Prevent running migrations in production environment
+            return;
+        }
+
         Schema::dropIfExists('password_resets');
     }
 };
