+++ database/migrations/2019_08_19_000000_create_failed_jobs_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\App;
 
 class CreateFailedJobsTable extends Migration
 {
@@ -13,6 +14,10 @@ class CreateFailedJobsTable extends Migration
      */
     public function up()
     {
+        if (App::environment() === 'production') {
+            // Prevent running migrations in production environment
+            return;
+        }
         Schema::create('failed_jobs', function (Blueprint $table) {
             $table->id();
             $table->string('uuid')->unique();
@@ -31,6 +36,10 @@ class CreateFailedJobsTable extends Migration
      */
     public function down()
     {
+        if (App::environment() === 'production') {
+            // Prevent running migrations in production environment
+            return;
+        }
         Schema::dropIfExists('failed_jobs');
     }
 }
