+++ database/migrations/2023_03_06_172611_create_categories_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\App;
 
 class CreateCategoriesTable extends Migration
 {
@@ -13,6 +14,11 @@ class CreateCategoriesTable extends Migration
      */
     public function up()
     {
+        if (App::environment() === 'production') {
+            // Prevent running migrations in production environment
+            return;
+        }
+
         Schema::create('categories', function (Blueprint $table) {
             $table->id();
             $table->string('name');
@@ -27,6 +33,11 @@ class CreateCategoriesTable extends Migration
      */
     public function down()
     {
+        if (App::environment() === 'production') {
+            // Prevent running migrations in production environment
+            return;
+        }
+
         Schema::dropIfExists('categories');
     }
 }
